Add required dependencies for game to launch
