+++ build.gradle
@@ -42,20 +42,30 @@ dependencies {
 	mappings "net.fabricmc:yarn:${project.yarn_mappings}:v2"
 	modImplementation "net.fabricmc:fabric-loader:${project.loader_version}"
 	modImplementation "net.fabricmc.fabric-api:fabric-api:${project.fabric_version}"
+	modImplementation "maven.modrinth:cloth-config:${project.clothconfig_version}"
 	// convenience and minor testing
 	modLocalRuntime "com.terraformersmc:modmenu:${modmenu_version}"
 	// wilder wild (for (s)hlogs)
-	modImplementation "maven.modrinth:wilder-wild:lZgIMgjb"
-	modImplementation "maven.modrinth:frozenlib:9y9515ej"
+	modImplementation "maven.modrinth:wilder-wild:${project.wilderwild_version}"
+	modImplementation "maven.modrinth:frozenlib:${project.frozenlib_version}"
+	// promenade
+	modImplementation "maven.modrinth:promenade:${project.promenade_version}"
+	modImplementation "maven.modrinth:dawn:${project.dawnapi_version}"
+	modImplementation "maven.modrinth:biolith:${project.biolith_version}"
+	modImplementation "com.terraformersmc.terraform-api:terraform-surfaces-api-v1:${project.terraform_surfaces_api_version}"
+	modImplementation "com.terraformersmc.terraform-api:terraform-shapes-api-v1:${terraform_shapes_version}"
+	modImplementation "com.terraformersmc.terraform-api:terraform-wood-api-v1:${terraform_wood_version}"
+	// terrablender
+	modImplementation "maven.modrinth:terrablender:${terrablender_version}"
+	implementation "com.electronwill.night-config:toml:${nightconfig_version}"
+	implementation "com.electronwill.night-config:core:${nightconfig_version}"
 	// all other mods
 	modImplementation "maven.modrinth:great-big-world:NIAb0XO6"
 	modImplementation "maven.modrinth:creo-lib:V8ENaaAV"
-	modImplementation "maven.modrinth:terrablender:CMDA8IcP"
 	//modImplementation "maven.modrinth:biome-makeover:rsoZ2Iqo"
 	modImplementation "maven.modrinth:architectury-api:Sbew3kXe"
 	modImplementation "maven.modrinth:cinderscapes:2mBOodyG"
 	modImplementation "maven.modrinth:gardens-of-the-dead:dwXQdKI3"
-	modImplementation "maven.modrinth:promenade:A5YD9uS3"
 	modImplementation "maven.modrinth:pearfection:KHR0uBvn"
 	modImplementation "maven.modrinth:hellions-sniffer+:DLjYM0KE"
 	modImplementation "maven.modrinth:traverse:9o4IEnHZ"
