+++ gradle.properties
@@ -14,3 +14,14 @@ archives_base_name=astound_compat
 # Dependencies
 fabric_version=0.91.0+1.20.1
 modmenu_version=7.2.2
+clothconfig_version=11.1.106+fabric
+wilderwild_version=2.1.14-mc1.20.1
+frozenlib_version=1.4.4-mc1.20.1
+promenade_version=v4.1.1
+dawnapi_version=v5.0.0
+biolith_version=1.0.0-alpha.9
+terraform_surfaces_api_version=8.0.1
+terraform_shapes_version=7.0.1
+terraform_wood_version=7.0.1
+terrablender_version=3.0.0.169
+nightconfig_version=3.6.4
